Initial Project Files for Craig Berntson's ASP.Net MVC Unit Testing Tutorial
I hope that headline is not too long. This is the package from Craig Berntson's
blog series on learning unit testing with ASP.Net MVC. The link is here -

http://www.craigberntson.com/blog/?p=1096

Even though MVC 5 is about to be released, it's probably impossible to work with
a tutorial that starts with that. Here is the project zip -

http://www.craigberntson.com/downloads/mvcunittesting1.zip