Initial Check-In for MvcMusicStore Tutorial
This is the MVC 4 lesson on the Microsoft site, http://bit.ly/15aGX6o . It
looks like I need to practice some MVC badly, esp. the data parts.

My attempt to try to practice unit testing failed. It may be whining, but the
dude who put the lesson together, Craig Berntson, provided a zip with the
MVC 4 project for his tutorial but neglected to include the DB parts, so the
code could not run. And I would have spent too much time trying to heal it. His
tutorial is at http://www.craigberntson.com/blog/?p=1096.