Initial Check-In of Guestbook
This is a training project from "ASP.NET MVC 4 in Action" by Jeff Palermo et al
from Manning. This is the project that starts in Chapter 2. I suppose it will
be a mandatory overview of MVC, but it will be the foundation for the next two
chapters so it needs to be completed.

I am starting to exclude the packages folder created by NuGet, since it can
always be rebuilt by Visual Studio.

It was interesting that they wanted the Unit Test project added. I wonder how
that will play out. I hope it will be good.