Completed Chapter 2 Guestbook
Added a Guestbook section with its own controller. The controller has a simple
Index action to list guestbook entries. It also has a Create action.

The chapter is structured a little weird. There is no indication that the app
should be run after you make changes at any point, so as you work through it
you'll find yourself deleting things you had just created a little while ago
for no reason. For example, in the Create POST action, we first return content
stating that it succeeded. Soon after that we replace it with a redirect to
the Index action. If we had executed the code to see the content return work
first then it would have meant something. Anyway, the lesson is we need to
build and execute after every change.